+++ src/CommandInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Leapt\GitWrapper;
+
+interface CommandInterface
+{
+    public function run(): string;
+}