+++ tests/Fixtures/CommandThatDoesNotImplementInterface.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Leapt\GitWrapper\Tests\Fixtures;
+
+final class CommandThatDoesNotImplementInterface
+{
+    public function run(): string
+    {
+        return '';
+    }
+}